BRS-1261 seed variances
